Moved all battle-related functionality of server into a new class
